Clean Up.

+++ Textures/CheckerTexture.cpp
@@ -10,7 +10,7 @@ CheckerTexture::CheckerTexture(Texture *t0, Texture *t1)
 
 Vector3 CheckerTexture::Value(const Vector2 &uv, const Vector3 &p) const
 {
-	float sines = sin(10 * p.x) * sin(10 * p.y) * sin(10 * p.z);
+	float sines = sinf(10 * p.x) * sinf(10 * p.y) * sinf(10 * p.z);
 
 	if (sines < 0)
 		return odd->Value(uv, p);
